uniProjects/vocabBot/requirements.txt
python-telegram-bot==21.0
requests==2.31.0
schedule==1.2.0
